+++ statick_tool/plugins/reporting/json_reporting_plugin.py
@@ -73,12 +73,18 @@ class JsonReportingPlugin(ReportingPlugin):
 
     def write_output(self, package: Package, level: str, line: str) -> bool:
         """Write JSON output to a file."""
-        if not self.plugin_context:
-            return False
+        # By default write report to the current directory.
+        output_dir = os.getcwd()
+        if (
+            self.plugin_context
+            and "output_directory" in self.plugin_context.args
+            and self.plugin_context.args.output_directory is not None
+        ):
+            # If an output directory is specified use it for the report.
+            output_dir = os.path.join(
+                self.plugin_context.args.output_directory, package.name + "-" + level
+            )
 
-        output_dir = os.path.join(
-            self.plugin_context.args.output_directory, package.name + "-" + level
-        )
         if not os.path.exists(output_dir):
             os.mkdir(output_dir)
         if not os.path.isdir(output_dir):