+++ CHANGELOG.md
@@ -71,6 +71,8 @@ max            |                | 223    | 127    | 64
   When running pytest a warning was given about this. The max-line-length
   was being set in the pytest section of our configuration. Moving it to
   the flake8 section fixes the warning and makes more sense logically. (#466, #469)
+- In the code climate and json reporting plugins write output files to current directory
+  if no output directory is specified. (#471)
 
 ## v0.9.2 - 2022-12-21